Update 201 entry-level-roles-in-tech.md
I think it's not in the current Curriculum workshop format (i.e. Design principles isn't something we use, table with duration, etc)) - look into some of the others on our workshop repo. But we can rejuggle that easily. >> Done, have rejuggled

The overall workshop should not be longer than 1.5 hours if I am not mistaken. Can we reduce it? >> Done as well


On the instruction I am missing if people should get together as a group and do this or each individual? Prep & Coursework is individual, course part 1 & 2 is in group

Is this search done in the class? Some of this yes, especially in the zoom in part

Could we instead of creating a table use post it notes and get people to work together on them (which is what you added in the end of the mind map)? Yes, this makes more sense indeed

Could they come up with the group categories (we can give it as hint)? Changed this bit as well

Can we move the "which one is for me" as a coursework (also known as backlog)? Done

Let me know if there are other feedbacks.